// apps/frontend/src/common/interfaces/patient/patient-workflow.interface.ts
export interface PaginationParams {
  page: number;
  limit: number;
}